WIP - main code compiles, but benchmarks do not fully compile.
Status: I am pulling through code incrementally from the db::iterator
but am hitting the wall when scan() gets instantiated in the
microbenchmark for OLC since iterator::get_key() and get_val() can not
be implemented without proper access to the various typedefs.

I am running into issues where the type information is not properly
visible for OLC.  This was fixed by a gist for unodb::db.  I need to
apply the same change concept in this branch to make additional
progress.

I think this needs to be addressed now.  The next thing after that is
to implement the OLC variants of the iterator methods.  And I can't do
that without access to the leaf and node types.  And that is precisely
what needs to be refactored.